refac(91026): Extrai task gerar_notificacao_pendencia_envio_prestacao_de_contas_async
